Add static cast for ember calls for span sizes ... we know we are less than 65K

+++ src/app/codegen-data-model/CodegenDataModel_Read.cpp
@@ -341,7 +341,7 @@ CHIP_ERROR CodegenDataModel::ReadAttribute(const InteractionModel::ReadAttribute
     record.clusterId                           = request.path.mClusterId;
     record.attributeId                         = request.path.mAttributeId;
     Protocols::InteractionModel::Status status = emAfReadOrWriteAttribute(
-        &record, &attributeMetadata, gEmberAttributeIOBufferSpan.data(), gEmberAttributeIOBufferSpan.size(),
+        &record, &attributeMetadata, gEmberAttributeIOBufferSpan.data(), static_cast<uint16_t>(gEmberAttributeIOBufferSpan.size()),
         /* write = */ false);
 
     if (status != Protocols::InteractionModel::Status::Success)
